test: merge the three mock_rest_client fixtures into one
acp/conftest.py, runtime/test_tools.py and runtime/test_contact_tools.py
each defined their own mock_rest_client fixture under the same name,
scaffolding overlapping agent_api_* namespaces independently.

Moves a single fixture to the root conftest.py as the union of what the
three already stubbed (chat/message/event creation, participant listing
and mutation, peer lookup, identity lookup), and deletes the three local
definitions. pytest resolves the fixture from the nearest conftest up
the directory tree, so every existing call site picks it up unchanged.

The ids each suite asserts on directly (room-new-123, msg-123, evt-123,
user-1, agent-2) are preserved exactly. The one behavior change ACP
inherits is a non-empty default participant list (previously empty) for
handle_prompt's mention-building path; nothing asserts on that path's
mention contents, and runtime/test_tools.py has three tests that already
depend on the non-empty default, so that default is the one kept.

Also drops the acp fixture's `client.agent_api_contacts = MagicMock()`
and contact_tools' `client.agent_api_contacts = MagicMock()` lines: both
reassigned an attribute to a value indistinguishable from what MagicMock
auto-vivifies, so they did nothing.
