+++ tests/runtime/test_tools.py
@@ -28,83 +28,6 @@ from band.runtime.tools import (
 
 
 @pytest.fixture
-def mock_rest_client():
-    """Mock AsyncRestClient for testing AgentTools."""
-    client = MagicMock()
-
-    # Mock create_agent_chat_message
-    message_response = MagicMock()
-    message_response.data = MagicMock()
-    message_response.data.model_dump.return_value = {
-        "id": "msg-123",
-        "content": "Hello",
-        "sender_id": "agent-1",
-    }
-    client.agent_api_messages.create_agent_chat_message = AsyncMock(
-        return_value=message_response
-    )
-
-    # Mock create_agent_chat_event
-    event_response = MagicMock()
-    event_response.data = MagicMock()
-    event_response.data.model_dump.return_value = {
-        "id": "evt-123",
-        "content": "Thinking...",
-        "message_type": "thought",
-    }
-    client.agent_api_events.create_agent_chat_event = AsyncMock(
-        return_value=event_response
-    )
-
-    # Mock list_agent_chat_participants
-    participant1 = make_participant_mock(
-        "user-1", "User One", "User", handle="user-one"
-    )
-    client.agent_api_participants.list_agent_chat_participants = AsyncMock(
-        return_value=MagicMock(data=[participant1])
-    )
-
-    # Mock list_agent_peers
-    peer1 = make_participant_mock(
-        "agent-2", "Agent Two", "Agent", handle="agent-two", description="Another agent"
-    )
-    peers_response = MagicMock()
-    peers_response.data = [peer1]
-    peers_response.metadata = MagicMock()
-    peers_response.metadata.page = 1
-    peers_response.metadata.page_size = 50
-    peers_response.metadata.total_count = 1
-    peers_response.metadata.total_pages = 1
-    peers_response.model_dump = MagicMock(
-        return_value={
-            "data": [
-                {
-                    "id": "agent-2",
-                    "name": "Agent Two",
-                    "type": "Agent",
-                    "description": "Another agent",
-                }
-            ],
-            "metadata": {
-                "page": 1,
-                "page_size": 50,
-                "total_count": 1,
-                "total_pages": 1,
-            },
-        }
-    )
-    client.agent_api_peers.list_agent_peers = AsyncMock(return_value=peers_response)
-
-    # Mock add_agent_chat_participant
-    client.agent_api_participants.add_agent_chat_participant = AsyncMock()
-
-    # Mock remove_agent_chat_participant
-    client.agent_api_participants.remove_agent_chat_participant = AsyncMock()
-
-    return client
-
-
-@pytest.fixture
 def participants():
     """Sample participants list."""
     return [
